Factored out some common reading/writing stuff into CStructUtil. HardDat writing works, and the following holds for the Binary instances as given in HardDat.hs:
x == decode (encode x)

However,

x == encode (decode x)

does not hold, because of junk bits that can be anything. In
particular, the original Dink executables just write the in-memory
representation to file, which means that the padding bytes and
the unused "hold" fields have whatever value was originally at
that memory location.